From PurchasedAndArrivedArticle to warehouseToWarehouse

+++ TomasGreen.Web/Areas/Import/Controllers/CompaniesController.cs
@@ -2,9 +2,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Localization;
 using TomasGreen.Model.Models;
 using TomasGreen.Web.BaseModels;
 using TomasGreen.Web.Data;
@@ -15,10 +17,13 @@ namespace TomasGreen.Web.Areas.Import.Controllers
     public class CompaniesController : BaseController
     {
         private readonly ApplicationDbContext _context;
-
-        public CompaniesController(ApplicationDbContext context)
+        private readonly IHostingEnvironment _hostingEnvironment;
+        private readonly IStringLocalizer<CompaniesController> _localizer;
+        public CompaniesController(ApplicationDbContext context, IHostingEnvironment hostingEnvironment, IStringLocalizer<CompaniesController> localizer)
         {
             _context = context;
+            _hostingEnvironment = hostingEnvironment;
+            _localizer = localizer;
         }
 
         // GET: Sales/Companies
@@ -56,10 +61,16 @@ namespace TomasGreen.Web.Areas.Import.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Name,Balance,Ruble,Discount,Purchases,SouldToUs,Paid,Received,LastBalance,LastBalanceDate,Locked,CreditReceived,CreditLimit")] Company company)
+        public async Task<IActionResult> Create(Company company)
         {
             if (ModelState.IsValid)
             {
+                if (NameAlreadyTaken(company))
+                {
+                    ModelState.AddModelError(nameof(company.Name), _localizer["Name is already taken."]);
+                    return View(company);
+                }
+
                 _context.Add(company);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
@@ -142,6 +153,16 @@ namespace TomasGreen.Web.Areas.Import.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var company = await _context.Companies.SingleOrDefaultAsync(m => m.ID == id);
+            if (company.IsOwner)
+            {
+                ViewBag.Error = _localizer["Couldn't delete. The Company is the owner."];
+                return View(company);
+            }
+            if (IsRelated(company))
+            {
+                ViewBag.Error = _localizer["Couldn't delete. The Post is already related to other entities."];
+                return View(company);
+            }
             _context.Companies.Remove(company);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
@@ -151,5 +172,16 @@ namespace TomasGreen.Web.Areas.Import.Controllers
         {
             return _context.Companies.Any(e => e.ID == id);
         }
+        #region
+        private bool NameAlreadyTaken(Company model)
+        {
+            return _context.Companies.Any(f => f.Name == model.Name && f.ID != model.ID);
+        }
+
+        private bool IsRelated(Company model)
+        {
+            return _context.ArticleWarehouseBalances.Any(a => a.CompanyID == model.ID);
+        }
+        #endregion
     }
 }
